+++ NEWS.md
@@ -1,5 +1,11 @@
 # News
 
+## [No version] No date
+
+### Changed
+
+* [mdl] Various bugfixes
+
 ## [v1.1.1] 2016-10-26
 
 ### Changed